plugins/requirements.txt
--find-links /usr/local/airflow/plugins
--no-index
--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.10.1/constraints-3.11.txt"
cx_Oracle
apache-airflow-providers-oracle
apache-airflow-providers-snowflake
apache-airflow-providers-jdbc
boto3
pandas
